Initialized Mobile App With Basic Features
This is the mobile counterpart for the ongoing bonaso-data-viewer project.

The idea behind this app is to give users the ability to perform basic functions offline and on
mobile (i.e., collecting responses in the field). Its not a full 'mobile port', rather a supplement.

This commit just has the basic react-native file structure and some SQLite integrations.