Have responses semi-working now.
Users can now fill out responses. A decent chunk of logic handling is there
and the submissions are coming out correctly (even though nothing is being done with them).

Need to fix the issue where a hidden question can still have values and also need to actually do
stuff with the submissions, but this isn't a terrible start, I hope.