Reworked offline/online statuses.
Big push this commit is that there is now a global connection context, and things that
require a connection to the server will only happen if there is an established connection.

Connection refreshes every 60 seconds.

Also did some minor file reorganization to clean up query writer, all those functions are now in the
sync-load-queries directory.